positionX and positionY from int to double.

+++ PongSpiel/src/de/demoncore/game/GameObject.java
@@ -8,11 +8,11 @@ public class GameObject {
 	public int groesseX;
 	public int groesseY;
 	
-	public int positionX;
-	public int positionY;
+	public double positionX;
+	public double positionY;
 	
 
-	public GameObject(int posX, int posY, int breite, int hoehe) {
+	public GameObject(double posX, double posY, int breite, int hoehe) {
 		positionX = posX;
 		positionY = posY;
 		groesseX = breite;
